requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==2.0.2
djangorestframework==3.7.7
future==0.16.0
gcloud==0.17.0
googleapis-common-protos==1.5.3
httplib2==0.10.3
idna==2.6
jws==0.1.3
line-bot-sdk==1.5.0
oauth2client==3.0.0
protobuf==3.5.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycryptodome==3.4.3
Pyrebase==3.0.27
python-dateutil==2.6.1
python-jwt==2.0.1
pytz==2018.3
requests==2.11.1
requests-toolbelt==0.7.0
rsa==3.4.2
six==1.11.0
urllib3==1.22
